+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-## [0.2.0-nullsafety.0] - 2020-12-03
+## [0.2.0] - 2021-08-11
 
 * Migrated to null safety.
+
+## [0.1.2] - 2021-08-11
+
+* Method `firstToLower()` - converts first character in the string to lower case.
+* Added `innim_lint` analysis ruleset.
+
+## [0.1.1] - 2021-04-07
+
+* Method `chunks()` - splits string by chunk with defined size.
+
 
 ## [0.1.0] - 2020-10-16